+++ extensions/vscode/src/types/messages/hostToWebviewMessages.ts
@@ -8,6 +8,8 @@ import {
   ContentRecordFile,
   PreContentRecord,
   ConfigurationError,
+  IntegrationRequest,
+  ServerSettings,
 } from "../../api";
 import { DeploymentSelector } from "../shared";
 
@@ -22,6 +24,8 @@ export enum HostToWebviewMessageType {
   PUBLISH_START = "publishStart",
   PUBLISH_FINISH_SUCCESS = "publishFinishSuccess",
   PUBLISH_FINISH_FAILURE = "publishFinishFailure",
+  CONTENT_RENDER_FINISHED = "contentRenderFinished",
+  CONTENT_RENDER_FAILURE = "contentRenderFailure",
   UPDATE_CONTENTRECORD_SELECTION = "updateContentRecordSelection",
   SAVE_SELECTION = "saveSelection",
   REFRESH_FILES = "refreshFiles",
@@ -31,6 +35,8 @@ export enum HostToWebviewMessageType {
   HIDE_DISABLE_OVERLAY = "hideDisableOverlay",
   SET_PATH_SEPARATOR = "setPathSeparator",
   UPDATE_SERVER_ENVIRONMENT = "updateServerEnvironment",
+  REFRESH_INTEGRATION_REQUESTS = "refreshIntegrationRequests",
+  REFRESH_SERVER_SETTINGS = "refreshServerSettings",
 }
 
 export type AnyHostToWebviewMessage<
@@ -53,6 +59,8 @@ export type HostToWebviewMessage =
   | PublishStartMsg
   | PublishFinishSuccessMsg
   | PublishFinishFailureMsg
+  | ContentRenderFinishedMsg
+  | ContentRenderFailureMsg
   | UpdateContentRecordSelectionMsg
   | SaveSelectionMsg
   | RefreshFilesMsg
@@ -61,7 +69,9 @@ export type HostToWebviewMessage =
   | ShowDisableOverlayMsg
   | HideDisableOverlayMsg
   | SetPathSeparatorMsg
-  | UpdateServerEnvironmentMsg;
+  | UpdateServerEnvironmentMsg
+  | RefreshIntegrationRequestsMsg
+  | RefreshServerSettingsMsg;
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function isHostToWebviewMessage(msg: any): msg is HostToWebviewMessage {
@@ -75,6 +85,8 @@ export function isHostToWebviewMessage(msg: any): msg is HostToWebviewMessage {
     msg.kind === HostToWebviewMessageType.PUBLISH_START ||
     msg.kind === HostToWebviewMessageType.PUBLISH_FINISH_SUCCESS ||
     msg.kind === HostToWebviewMessageType.PUBLISH_FINISH_FAILURE ||
+    msg.kind === HostToWebviewMessageType.CONTENT_RENDER_FINISHED ||
+    msg.kind === HostToWebviewMessageType.CONTENT_RENDER_FAILURE ||
     msg.kind === HostToWebviewMessageType.UPDATE_CONTENTRECORD_SELECTION ||
     msg.kind === HostToWebviewMessageType.SAVE_SELECTION ||
     msg.kind === HostToWebviewMessageType.REFRESH_FILES ||
@@ -83,7 +95,9 @@ export function isHostToWebviewMessage(msg: any): msg is HostToWebviewMessage {
     msg.kind === HostToWebviewMessageType.SHOW_DISABLE_OVERLAY ||
     msg.kind === HostToWebviewMessageType.HIDE_DISABLE_OVERLAY ||
     msg.kind === HostToWebviewMessageType.SET_PATH_SEPARATOR ||
-    msg.kind === HostToWebviewMessageType.UPDATE_SERVER_ENVIRONMENT
+    msg.kind === HostToWebviewMessageType.UPDATE_SERVER_ENVIRONMENT ||
+    msg.kind === HostToWebviewMessageType.REFRESH_INTEGRATION_REQUESTS ||
+    msg.kind === HostToWebviewMessageType.REFRESH_SERVER_SETTINGS
   );
 }
 
@@ -126,6 +140,10 @@ export type PublishFinishFailureMsg = AnyHostToWebviewMessage<
     };
   }
 >;
+export type ContentRenderFinishedMsg =
+  AnyHostToWebviewMessage<HostToWebviewMessageType.CONTENT_RENDER_FINISHED>;
+export type ContentRenderFailureMsg =
+  AnyHostToWebviewMessage<HostToWebviewMessageType.CONTENT_RENDER_FAILURE>;
 export type UpdateContentRecordSelectionMsg = AnyHostToWebviewMessage<
   HostToWebviewMessageType.UPDATE_CONTENTRECORD_SELECTION,
   {
@@ -181,5 +199,19 @@ export type UpdateServerEnvironmentMsg = AnyHostToWebviewMessage<
   HostToWebviewMessageType.UPDATE_SERVER_ENVIRONMENT,
   {
     environment: string[];
+  }
+>;
+
+export type RefreshIntegrationRequestsMsg = AnyHostToWebviewMessage<
+  HostToWebviewMessageType.REFRESH_INTEGRATION_REQUESTS,
+  {
+    integrationRequests: IntegrationRequest[];
+  }
+>;
+
+export type RefreshServerSettingsMsg = AnyHostToWebviewMessage<
+  HostToWebviewMessageType.REFRESH_SERVER_SETTINGS,
+  {
+    serverSettings: ServerSettings;
   }
 >;