+++ extensions/vscode/src/utils/url.ts
@@ -1,4 +1,7 @@
-// Copyright (C) 2024 by Posit Software, PBC.
+// Copyright (C) 2025 by Posit Software, PBC.
+
+import config from "../../src/config";
+import { extractGUID } from "./guid";
 
 export const formatURL = (input: string): string => {
   // check if the URL starts with a scheme
@@ -15,4 +18,52 @@ export const normalizeURL = (input: string): string => {
     result += "/";
   }
   return result;
+};
+
+const isMatchingURL = (input: string, regex: RegExp) => {
+  const guid = extractGUID(input)?.at(0);
+  const guidlessInput = guid ? input.split(guid)[0] : "";
+  return regex.test(guidlessInput);
+};
+
+// All the following inputs will return "true" (production environment):
+//   "https://connect.posit.cloud/my-account/content/adffa505-08c7-450f-88d0-f42957f56eff"
+//   "https://connect.posit.cloud/user-profile-123/content/adffa505-08c7-450f-88d0-f42957f56eff"
+//   "https://connect.posit.cloud/slug123/content/adffa505-08c7-450f-88d0-f42957f56eff"
+// All the following inputs will return "false":
+//   "https://connect.posit.cloud/slug_with_underscore/content/adffa505-08c7-450f-88d0-f42957f56eff" // underscores
+//   "https://connect.posit.cloud/my account/content/adffa505-08c7-450f-88d0-f42957f56eff" // space
+//   "https://connect.posit.cloud/my-account/folder/content/adffa505-08c7-450f-88d0-f42957f56eff" // extra path segment
+export const isConnectCloudContentURL = (input: string) => {
+  const regex = new RegExp(
+    `^${config.connectCloudURL}/[a-zA-Z0-9-]+/content/$`,
+  );
+  return isMatchingURL(input, regex);
+};
+
+// All the following inputs will return "true":
+//   "https://connect.company.co/connect/#/apps/adffa505-08c7-450f-88d0-f42957f56eff";
+//   "https://connect.my_sub.company.com/connect/#/apps/adffa505-08c7-450f-88d0-f42957f56eff";
+//   "https://connect.another.deep_sub.domain_name.org/connect/#/apps/adffa505-08c7-450f-88d0-f42957f56eff";
+//   "https://connect.only-one.co/connect/#/apps/adffa505-08c7-450f-88d0-f42957f56eff";
+//   "https://company.co/data-science/2025/staging-server/#/apps/adffa505-08c7-450f-88d0-f42957f56eff";
+// All the following inputs will return "false":
+//   "https://connect.company.co/wrong/connect/#/path/apps/adffa505-08c7-450f-88d0-f42957f56eff"; // wrong path
+//   "https://connect.company.co/connect/#apps/adffa505-08c7-450f-88d0-f42957f56eff"; // missing slash
+//   "https://connect.invalid.com/connect/#/@pps/adffa505-08c7-450f-88d0-f42957f56eff"; // invalid character
+export const isConnectContentURL = (input: string) => {
+  const regex = /^http(s?):\/\/.+\/#\/apps\/$/;
+  return isMatchingURL(input, regex);
+};
+
+export const extractConnectCloudAccount = (input: string) => {
+  // sample URL: "https://connect.posit.cloud/my-account/content/adffa505-08c7-450f-88d0-f42957f56eff"
+  const url = new URL(input);
+
+  // the pathname is the part of the URL after the domain, e.g., "/my-account/content/adffa505-08c7-450f-88d0-f42957f56eff"
+  const pathSegments = url.pathname.split("/");
+
+  // for the given URL, the array will be ["", "my-account", "content", "adffa505-08c7-450f-88d0-f42957f56eff"]
+  // the account part is at index 1 of the array
+  return pathSegments.at(1);
 };