+++ extensions/vscode/webviews/homeView/src/stores/home.ts
@@ -11,8 +11,12 @@ import {
   ConfigurationError,
   isPreContentRecord,
   ServerType,
+  ServerSettings,
 } from "../../../../src/api";
-import { isConfigurationError } from "../../../../src/api/types/configurations";
+import {
+  IntegrationRequest,
+  isConfigurationError,
+} from "../../../../src/api/types/configurations";
 import { WebviewToHostMessageType } from "../../../../src/types/messages/webviewToHostMessages";
 import { RPackage } from "../../../../src/api/types/packages";
 import { DeploymentSelector } from "../../../../src/types/shared";
@@ -30,6 +34,8 @@ export const useHomeStore = defineStore("home", () => {
   const platformFileSeparator = ref<string>("/");
   const publishInProgress = ref(false);
   const publishInitiated = ref(false);
+  const contentRenderInProgress = ref(false);
+  const contentRenderFailed = ref(false);
 
   const contentRecords = ref<(ContentRecord | PreContentRecord)[]>([]);
   const configurations = ref<Configuration[]>([]);
@@ -39,8 +45,12 @@ export const useHomeStore = defineStore("home", () => {
     return credentials.value.sort((a, b) => a.name.localeCompare(b.name));
   });
 
+  const serverSettings = ref<ServerSettings | undefined>(undefined);
+
   const serverSecrets = ref<Set<string>>(new Set());
   const secrets = ref(new Map<string, string | undefined>());
+
+  const integrationRequests = ref<IntegrationRequest[]>([]);
 
   const environment = computed((): Map<string, string> => {
     const result = new Map<string, string>();
@@ -215,10 +225,9 @@ export const useHomeStore = defineStore("home", () => {
     });
   };
 
-  watch([selectedContentRecord], () => {
-    updateSelectionIsPreContentRecordState();
-    updateSelectionIsConnectContentRecordState();
-  });
+  watch([selectedContentRecord], () =>
+    updateSelectionIsPreContentRecordState(),
+  );
 
   const updateSelectionIsPreContentRecordState = () => {
     const hostConduit = useHostConduitService();
@@ -228,19 +237,6 @@ export const useHomeStore = defineStore("home", () => {
         state: isPreContentRecord(selectedContentRecord.value)
           ? "true"
           : "false",
-      },
-    });
-  };
-
-  const updateSelectionIsConnectContentRecordState = () => {
-    const hostConduit = useHostConduitService();
-    const serverType =
-      selectedContentRecord.value?.serverType || ServerType.CONNECT;
-    const productType = getProductType(serverType);
-    hostConduit.sendMsg({
-      kind: WebviewToHostMessageType.UPDATE_SELECTION_IS_CONNECT_CONTENT_RECORD,
-      content: {
-        state: isConnectProduct(productType) ? "true" : "false",
       },
     });
   };
@@ -378,6 +374,13 @@ export const useHomeStore = defineStore("home", () => {
     },
   };
 
+  const isConnectCloud = computed((): boolean => {
+    const serverType =
+      selectedContentRecord.value?.serverType || ServerType.CONNECT;
+    const productType = getProductType(serverType);
+    return isConnectCloudProduct(productType);
+  });
+
   const anyActiveAlerts = computed(() => {
     return (
       !config.active.isAlertActive.value &&
@@ -465,18 +468,23 @@ export const useHomeStore = defineStore("home", () => {
     showDisabledOverlay,
     publishInProgress,
     publishInitiated,
+    contentRenderInProgress,
+    contentRenderFailed,
     contentRecords,
     configurations,
     configurationsInError,
     credentials,
     sortedCredentials,
+    serverSettings,
     serverSecrets,
     secrets,
+    integrationRequests,
     environment,
     duplicatedEnvironmentVariables,
     selectedContentRecord,
     selectedConfiguration,
     serverCredential,
+    isConnectCloud,
     initializingRequestComplete,
     lastContentRecordResult,
     lastContentRecordMsg,